refactor ventilations
